Added Color class.
